Refactored code to use vector<char> instead of strings for messages
Strings don't seem to work with binary files, like images. Things
like '\0' are incorrectly represented with strings. Currently
working on figuring out what the problem is with large file transfers.
Most likely an issue with buffer size or the client receiving too
slowly for some reason, so the server is blocked indefinitely.

Next steps would be:
 1) to do some more error checking on the server's side for requests
    with bad headers
 2) implement select() asynchronous concurrency
 3) HTTP timeout